Refactor db:seeds (#14159)
Resolves #13791 

### Description

* Split `db/seeds.rb` into separate classes
* add tests for seeds
* add CircleCI job to test demo features (seeds, assets)

### Acceptance Criteria
- [x] Code compiles correctly
