+++ lib/tasks/local/vacols.rake
@@ -3,6 +3,8 @@
 require "csv"
 require "rainbow"
 require_relative "../../helpers/vacols_csv_reader"
+require_relative "../../../db/seeds/base"
+require_relative "../../../db/seeds/facols"
 
 namespace :local do
   namespace :vacols do
@@ -33,18 +35,7 @@ namespace :local do
 
     desc "Seeds local VACOLS"
     task seed: :environment do
-      date_shift = Time.now.utc.beginning_of_day - Time.utc(2017, 12, 10)
-      hearing_date_shift = Time.now.utc.beginning_of_day - Time.utc(2017, 7, 25)
-
-      vacols_models.each do |model|
-        if model == VACOLS::CaseHearing
-          VacolsCSVReader.new(model, hearing_date_shift).call
-        else
-          VacolsCSVReader.new(model, date_shift).call
-        end
-      end
-
-      create_issrefs
+      Seeds::Facols.new.local_vacols_seed!
     end
 
     # Do not check in the result of running this without talking with Chris. We need to certify that there
@@ -117,68 +108,6 @@ namespace :local do
     end
 
     private
-
-    def vacols_models
-      [
-        VACOLS::Case,
-        VACOLS::Folder,
-        VACOLS::Representative,
-        VACOLS::Correspondent,
-        VACOLS::CaseIssue,
-        VACOLS::Note,
-        VACOLS::CaseHearing,
-        VACOLS::Actcode,
-        VACOLS::Decass,
-        VACOLS::Staff,
-        VACOLS::Vftypes,
-        VACOLS::Issref,
-        VACOLS::TravelBoardSchedule
-      ]
-    end
-
-    # rubocop:disable Metrics/MethodLength
-    def create_issrefs
-      # creates VACOLS::Issrefs added later than our sanitized UAT copy
-      fiduciary_issue = {
-        prog_code: "12",
-        prog_desc: "Fiduciary"
-      }
-
-      # Issref_dump.csv has 1 Fiduciary issue
-      return if VACOLS::Issref.where(**fiduciary_issue).count > 1
-
-      FactoryBot.create(
-        :issref,
-        **fiduciary_issue,
-        iss_code: "01",
-        iss_desc: "Fiduciary Appointment"
-      )
-      FactoryBot.create(
-        :issref,
-        **fiduciary_issue,
-        iss_code: "02",
-        iss_desc: "Hub Manager removal of a fiduciary under 13.500"
-      )
-      FactoryBot.create(
-        :issref,
-        **fiduciary_issue,
-        iss_code: "03",
-        iss_desc: "Hub Manager misuse determination under 13.400"
-      )
-      FactoryBot.create(
-        :issref,
-        **fiduciary_issue,
-        iss_code: "04",
-        iss_desc: "RO Director decision upon recon of a misuse determ"
-      )
-      FactoryBot.create(
-        :issref,
-        **fiduciary_issue,
-        iss_code: "05",
-        iss_desc: "Dir of PFS negligence determination for reissuance"
-      )
-    end
-    # rubocop:enable Metrics/MethodLength
 
     def bgs_record_from_case(cases, case_descriptors)
       CSV.open(Rails.root.join("local/vacols", "bgs_setup.csv"), "wb") do |csv|
